requirements.txt
Flask>=2.0
python-dotenv
Flask-MySQLdb
Flask-SQLAlchemy
Flask-Migrate
mysqlclient
google-auth
google-auth-oauthlib
google-api-python-client
werkzeug
